Remove unnecessary panic
This is a breaking change as it changes the signature of prometheus
factory but it removes unnecessary panic in the code. Previously the
function changed was defined with variadic parameter for additional
list of metrics just to make the additional list optional. This can be
done also by using a variadic parameter where each metric is passed on
its own. This should be also preferred as user is not forced to create a
slice when passing a parameter.
We also use the existing lengths to pre-calculate the capacity of the
final list of metrics.