feat: Add DisableBodyReading option to fix performance issue (#66)
* feat: Add DisableBodyReading option to fix performance issue

This commit introduces a new 'DisableBodyReading' option in the config
to prevent the middleware from reading the request body when calculating
the request size. This addresses the performance degradation reported
in issue #64 by avoiding the expensive operation of reading the entire body.

When 'DisableBodyReading' is set to true, the middleware will fall back
to using the 'ContentLength' header to determine the request size.

The README has been updated to document this new option.

Fixes #64

* test: Add backward compatibility test for NewPrometheus